Moved activity card into its own widget. This was need to make tests to ensure they were created
